[refactor] main 패키지 로 이동
