+++ src/main/java/lass9436/chess/Rank.java
@@ -1,10 +1,10 @@
-package chess;
+package lass9436.chess;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import chess.pieces.Piece;
+import lass9436.chess.pieces.Piece;
 
 public class Rank {
