+++ src/main/java/lass9436/chess/pieces/NoPiece.java
@@ -1,8 +1,8 @@
-package chess.pieces;
+package lass9436.chess.pieces;
 
 import java.util.List;
 
-import chess.Position;
+import lass9436.chess.Position;
 
 public class NoPiece extends Piece {
 	public NoPiece(Piece.Type type, Piece.Color color, Piece.Representation representation) {
