+++ src/main/java/lass9436/chess/pieces/Pawn.java
@@ -1,9 +1,9 @@
-package chess.pieces;
+package lass9436.chess.pieces;
 
 import java.util.ArrayList;
 import java.util.List;
 
-import chess.Position;
+import lass9436.chess.Position;
 
 public class Pawn extends Piece {
 	public Pawn(Type type, Color color, Representation representation) {
